增加about页面+修改画3D图片
增加about页面+修改画3D图片

+++ GrapeFarm/Home/Commit3DDataVC.h
@@ -11,8 +11,8 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @interface Commit3DDataVC : BaseVC
-@property (nonatomic, assign) CGFloat max_r;
-@property (nonatomic, assign) CGFloat mix_r;
+@property (nonatomic, assign) NSInteger max_r;
+@property (nonatomic, assign) NSInteger mix_r;
 @property (nonatomic, strong) NSArray *circles;
 @property (nonatomic, copy)NSString *modelString;
 @end
